geo-agri-analyst/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
basicsr==1.4.2
scikit-image==0.22.0
numpy==1.24.3
python-multipart==0.0.6
aiohttp==3.9.0
Pillow==10.1.0
opencv-python==4.8.1.78